backend/requirements.txt
pywebview
pyinstaller
bs4
win10toast